requirements.txt
streamlit
python-dotenv
langchain-core
langchain-community
langchain-openai
langchain-groq
mysql-connector-python
urllib3